Move peer store utils to the peer package
