+++ commands/peers/deletepeer.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 
 	"github.com/gluster/glusterd2/client"
-	"github.com/gluster/glusterd2/context"
+	"github.com/gluster/glusterd2/peer"
 
 	"github.com/gorilla/mux"
 )
@@ -18,12 +18,12 @@ func deletePeerHandler(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	if !context.Store.PeerExists(id) {
+	if !peer.Exists(id) {
 		client.SendResponse(w, -1, http.StatusNotFound, "", http.StatusNotFound, nil)
 		return
 	}
 
-	if e := context.Store.DeletePeer(id); e != nil {
+	if e := peer.DeletePeer(id); e != nil {
 		client.SendResponse(w, -1, http.StatusInternalServerError, e.Error(), http.StatusInternalServerError, nil)
 	} else {
 		client.SendResponse(w, 0, 0, "", http.StatusNoContent, nil)
