+++ commands/peers/getpeer.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 
 	"github.com/gluster/glusterd2/client"
-	"github.com/gluster/glusterd2/context"
+	"github.com/gluster/glusterd2/peer"
 
 	"github.com/gorilla/mux"
 )
@@ -18,7 +18,7 @@ func getPeerHandler(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	if peer, err := context.Store.GetPeer(id); err != nil {
+	if peer, err := peer.GetPeer(id); err != nil {
 		client.SendResponse(w, -1, http.StatusNotFound, err.Error(), http.StatusNotFound, "")
 	} else {
 		client.SendResponse(w, 0, 0, "", http.StatusOK, peer)
