+++ peer/store-utils.go
@@ -1,4 +1,4 @@
-package store
+package peer
 
 // This file contains helper functions facilitate easier interaction with the
 // peer information stores in the store
@@ -6,21 +6,22 @@ package store
 import (
 	"encoding/json"
 
-	"github.com/gluster/glusterd2/peer"
+	"github.com/gluster/glusterd2/context"
+	"github.com/gluster/glusterd2/store"
 
 	log "github.com/Sirupsen/logrus"
 )
 
 const (
-	peerPrefix string = glusterPrefix + "peers/"
+	peerPrefix string = store.GlusterPrefix + "peers/"
 )
 
 func init() {
-	prefixes = append(prefixes, peerPrefix)
+	context.Store.InitPrefix(peerPrefix)
 }
 
 // AddOrUpdatePeer adds/updates given peer in the store
-func (s *GDStore) AddOrUpdatePeer(p *peer.Peer) error {
+func AddOrUpdatePeer(p *Peer) error {
 	json, err := json.Marshal(p)
 	if err != nil {
 		return err
@@ -28,7 +29,7 @@ func (s *GDStore) AddOrUpdatePeer(p *peer.Peer) error {
 
 	idStr := p.ID.String()
 
-	if err := s.Put(peerPrefix+idStr, json, nil); err != nil {
+	if err := context.Store.Put(peerPrefix+idStr, json, nil); err != nil {
 		return err
 	}
 
@@ -36,13 +37,13 @@ func (s *GDStore) AddOrUpdatePeer(p *peer.Peer) error {
 }
 
 // GetPeer returns specified peer from the store
-func (s *GDStore) GetPeer(id string) (*peer.Peer, error) {
-	pair, err := s.Get(peerPrefix + id)
+func GetPeer(id string) (*Peer, error) {
+	pair, err := context.Store.Get(peerPrefix + id)
 	if err != nil || pair == nil {
 		return nil, err
 	}
 
-	var p peer.Peer
+	var p Peer
 	if err := json.Unmarshal(pair.Value, &p); err != nil {
 		return nil, err
 	}
@@ -50,16 +51,16 @@ func (s *GDStore) GetPeer(id string) (*peer.Peer, error) {
 }
 
 // GetPeers returns all available peers in the store
-func (s *GDStore) GetPeers() ([]peer.Peer, error) {
-	pairs, err := s.List(peerPrefix)
+func GetPeers() ([]Peer, error) {
+	pairs, err := context.Store.List(peerPrefix)
 	if err != nil || pairs == nil {
 		return nil, err
 	}
 
-	peers := make([]peer.Peer, len(pairs))
+	peers := make([]Peer, len(pairs))
 
 	for i, pair := range pairs {
-		var p peer.Peer
+		var p Peer
 
 		if err := json.Unmarshal(pair.Value, &p); err != nil {
 			log.WithFields(log.Fields{
@@ -75,13 +76,13 @@ func (s *GDStore) GetPeers() ([]peer.Peer, error) {
 }
 
 // DeletePeer deletes given peer from the store
-func (s *GDStore) DeletePeer(id string) error {
-	return s.Delete(peerPrefix + id)
+func DeletePeer(id string) error {
+	return context.Store.Delete(peerPrefix + id)
 }
 
-// PeerExists checks if given peer is present in the store
-func (s *GDStore) PeerExists(id string) bool {
-	b, e := s.Exists(peerPrefix + id)
+// Exists checks if given peer is present in the store
+func Exists(id string) bool {
+	b, e := context.Store.Exists(peerPrefix + id)
 	if e != nil {
 		return false
 	}
